requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic
httpx
python-dotenv
loguru
sentry-sdk
pytest
pytest-asyncio
httpx[http2]
pydantic[email]
python-multipart
bcrypt==3.2.2